Double led controllers

+++ code/NWL2022/artnet_controller/index.js
@@ -1,10 +1,16 @@
-var options = {
+var options_left = {
   host: "192.168.50.50",
 
   sendAll: true,
 };
+var options_right = {
+  host: "192.168.50.51",
 
-var artnet = require("artnet")(options);
+  sendAll: true,
+};
+
+var artnet = require("artnet")(options_left);
+var artnet_right = require("artnet")(options_right);
 var nodeCleanup = require("node-cleanup");
 
 var start_channel = 0;
@@ -14,6 +20,7 @@ var num_leds_x = 5;
 var color = [255, 0, 0];
 var num_channels_total = num_leds_x * num_leds_y * channels_per_led * 2;
 var leds = new Uint8Array(num_channels_total);
+var leds_right = new Uint8Array(num_channels_total);
 for (let i = 0; i < num_channels_total; i += 3) {
   leds[i] = 255;
   leds[i + 1] = 255;
@@ -105,17 +112,33 @@ var ledLoop = setInterval(function () {
   // }
   rows_to_led_values();
   artnet.set(1, leds);
+  artnet_right.set(1, leds_right);
 }, 1000 / 30);
 
 // Graceful cleanup
+var numArtnetsCleanedUp = 0;
 nodeCleanup(function (exitCode, signal) {
   if (signal) {
     console.log("Shutting down");
     clearInterval(ledLoop);
     artnet.set(1, new Uint8Array(num_channels_total), function (err, red) {
       artnet.close();
-      process.kill(process.pid, signal);
+      numArtnetsCleanedUp += 1;
+      if (numArtnetsCleanedUp == 2) {
+        process.kill(process.pid, signal);
+      }
     });
+    artnet_right.set(
+      1,
+      new Uint8Array(num_channels_total),
+      function (err, red) {
+        artnet_right.close();
+        numArtnetsCleanedUp += 1;
+        if (numArtnetsCleanedUp == 2) {
+          process.kill(process.pid, signal);
+        }
+      }
+    );
     nodeCleanup.uninstall(); // don't call cleanup handler again
     return false;
   }
@@ -130,6 +153,11 @@ function rows_to_led_values() {
     leds[i + 1] = default_color.g;
     leds[i + 2] = default_color.b;
   }
+  for (let i = 0; i < num_channels_total; i += 3) {
+    leds_right[i] = default_color.r;
+    leds_right[i + 1] = default_color.g;
+    leds_right[i + 2] = default_color.b;
+  }
   for (let i = 0; i < row_leds.length; i++) {
     for (let x = 0; x < row_leds[i].num_leds; x++) {
       // if (x >= num_leds_x) {
@@ -140,6 +168,9 @@ function rows_to_led_values() {
       leds[index] = row_leds[i].left_color.r;
       leds[index + 1] = row_leds[i].left_color.g;
       leds[index + 2] = row_leds[i].left_color.b;
+      leds_right[index] = row_leds[i].right_color.r;
+      leds_right[index + 1] = row_leds[i].right_color.g;
+      leds_right[index + 2] = row_leds[i].right_color.b;
     }
     for (let x = row_leds[i].num_leds; x < num_leds_x; x++) {
       let index = xy_to_index(x, i);
@@ -147,6 +178,9 @@ function rows_to_led_values() {
       leds[index] = default_color.r;
       leds[index + 1] = default_color.g;
       leds[index + 2] = default_color.b;
+      leds_right[index] = default_color.r;
+      leds_right[index + 1] = default_color.g;
+      leds_right[index + 2] = default_color.b;
     }
   }
 }
@@ -163,7 +197,7 @@ var udpPort = new osc.UDPPort({
 // Listen for incoming OSC messages.
 udpPort.on("message", function (oscMsg, timeTag, info) {
   // console.log("An OSC message just arrived!", oscMsg);
-  console.log(oscMsg.args);
+  // console.log(oscMsg.args);
   let num_leds = oscMsg.args[0].value;
   let left_color = oscMsg.args[1].value;
   let right_color = oscMsg.args[2].value;
